metadata: move UntranslatedPeer to address_translator.rs
`UntranslatedPeer` is only used for address translation purposes
and is not part of cluster metadata tables, so it makes sense for it
to reside alongside `AddressTranslator`.
