+++ scylla/src/cloud/config.rs
@@ -9,11 +9,10 @@ use tracing::warn;
 use uuid::Uuid;
 
 use crate::client::session::TlsContext;
-use crate::cluster::metadata::UntranslatedPeer;
 use crate::cluster::node::resolve_hostname;
 use crate::errors::TranslationError;
 use crate::network::{TlsConfig, TlsError};
-use crate::policies::address_translator::AddressTranslator;
+use crate::policies::address_translator::{AddressTranslator, UntranslatedPeer};
 
 #[non_exhaustive]
 #[derive(Debug, Error)]
